pyproject.toml
[project]
name = "scotland-yard"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "faker>=37.1.0",
    "fastapi[standard]>=0.115.12",
    "pydantic-settings>=2.9.1",
    "pyjwt[crypto]>=2.10.1",
    "tortoise-cli>=0.1.2",
    "tortoise-orm[asyncpg]>=0.25.0",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "poethepoet>=0.34.0",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
    "pytest-env>=1.1.5",
    "pytest-mock>=3.14.0",
    "pytest-randomly>=3.16.0",
]

[tool.poe.tasks.dev-setup]
sequence = [
    "uv sync",
    "pre-commit install",
]
help = "set up development environment"
default_item_type = "cmd"

[tool.poe.tasks.lint]
cmd = "pre-commit run --all-files"
help = "execute linters"

[tool.poe.tasks.dev]
cmd = """
docker compose
    -f docker-compose.yml
    -f docker-compose.dev.yml
    watch
"""
help = "start development server"

[tool.poe.tasks.pyshell]
cmd = "docker compose exec -it app tortoise-cli -c app.settings.TORTOISE_ORM shell"
help = "start interactive Python shell"

[tool.poe.tasks.dbshell]
cmd = "docker compose exec -it postgres bash -c 'psql -U $POSTGRES_USER -h localhost'"
help = "start interactive database shell"

[tool.poe.tasks.test]
cmd = """
docker compose
    -f docker-compose.yml
    -f docker-compose.test.yml
    up
    --abort-on-container-exit
    --exit-code-from app
    --build
"""
help = "start development server"

[tool.ruff]
cache-dir = ".cache/ruff"

[tool.ruff.lint]
select = [
    "F", # Pyflakes
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "I", # isort
    "PGH", # pygrep-hooks
    "UP", # pyupgrade
    "S", # flake8-bandit
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "T20", # flake8-print
    "SIM", # flake8-simplify
    "ARG", # flake8-unused-arguments
    "FAST", # FastAPI
    "PTH", # flake8-use-pathlib
    "FURB", # refurb
    "RUF", # Ruff-specific rules
]
ignore = [
    # Suppressed by ruff formatter.
    "W191", # tab-indentation
    "E111", # indentation-with-invalid-multiple
    "E114", # indentation-with-invalid-multiple-comment
    "E117", # over-indented
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101", # assert
    "S105", # hardcoded-password-string
]

[tool.mypy]
strict = true
cache_dir = ".cache/mypy"

[tool.pytest.ini_options]
pythonpath = "."
cache_dir = ".cache/pytest"
addopts = [
    "--doctest-modules",
    "--import-mode", "importlib",
    "--cov-config", "pyproject.toml",
    "--cov-report", "term",
    "--cov-report", "html",
    "--cov-fail-under", "100",
]
markers = [
    "only: shorthand mark used to select a small subset of test cases for debugging",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.pytest_env]
TESTING = "true"

[tool.coverage.run]
branch = true
data_file = ".cache/coverage/data.db"
disable_warnings = ["no-data-collected"]

[tool.coverage.report]
exclude_also = [
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@overload",
]

[tool.coverage.html]
directory = ".cache/cov-report"
